Updated Sessions with RecyclerView
Added a recycler view to the sessions page, which means I added the corresponding data class, view adapter class, and session_layout.  An event table model was added to handle database interactions.  Icons were added for the rsvp section of the list.  Minor fixes for navigation buttons and just display across the board.